# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: DUODROP Music Streaming Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/{userId}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          description: User not found

    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user profile
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /artists/{artistId}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get artist profile
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
        "404":
          description: Not found

    patch:
      operationId: updateArtist
      tags: [artists]
      summary: Update artist profile
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Artist"
  /artists/{artistId}/songs:
    get:
      operationId: getArtistSongs
      tags: [artists]
      summary: Get songs by artist
      parameters:
        - name: artistId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"

  # SONGS
  /songs:
    get:
      operationId: listSongs
      tags: [songs]
      summary: List songs
      parameters:
        - name: genre
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"

    post:
      operationId: createSong
      tags: [songs]
      summary: Upload a new song
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongInput"
      responses:
        "201":
          description: Song created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
  /songs/{songId}:
    get:
      operationId: getSong
      tags: [songs]
      summary: Get song by ID
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Song found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"
        "404":
          description: Not found

    patch:
      operationId: updateSong
      tags: [songs]
      summary: Update song metadata
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SongUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Song"

    delete:
      operationId: deleteSong
      tags: [songs]
      summary: Delete a song
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
  /songs/{songId}/play:
    post:
      operationId: playSong
      tags: [songs]
      summary: Record a song play
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Play recorded
  /songs/{songId}/like:
    post:
      operationId: likeSong
      tags: [songs]
      summary: Like a song
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Liked

    delete:
      operationId: unlikeSong
      tags: [songs]
      summary: Unlike a song
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Unliked
  /songs/{songId}/comments:
    get:
      operationId: getSongComments
      tags: [songs]
      summary: Get comments for a song
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"

    post:
      operationId: addComment
      tags: [songs]
      summary: Add a comment
      parameters:
        - name: songId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Comment added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # DISCOVER
...
  /admin/songs:
    get:
      operationId: listAdminSongs
      tags: [admin]
      summary: List all songs (admin)
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: All songs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Song"
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, email, role, createdAt]
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [fan, artist, admin]
        bio:
          type: ["string", "null"]
        profilePicUrl:
          type: ["string", "null"]
        followersCount:
          type: integer
        followingCount:
          type: integer
        isFollowing:
          type: ["boolean", "null"]
        themePreference:
          type: string
          enum: [dark, light]
        createdAt:
          type: string
    UserUpdate:
      type: object
      properties:
        username:
          type: string
        bio:
          type: ["string", "null"]
        profilePicUrl:
          type: ["string", "null"]
        themePreference:
          type: string
          enum: [dark, light]
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
    SocialLinks:
      type: object
      properties:
        instagram:
          type: ["string", "null"]
        twitter:
          type: ["string", "null"]
        facebook:
          type: ["string", "null"]
        youtube:
          type: ["string", "null"]

    # ARTIST
    Artist:
      type: object
      required: [id, username, followersCount, songsCount]
      properties:
        id:
          type: integer
        username:
          type: string
        bio:
          type: ["string", "null"]
        profilePicUrl:
          type: ["string", "null"]
        followersCount:
          type: integer
        songsCount:
          type: integer
        totalPlays:
          type: integer
        totalEarnings:
          type: number
        isFollowing:
          type: ["boolean", "null"]
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"
        isEligibleForEarnings:
          type: boolean
    ArtistUpdate:
      type: object
      properties:
        bio:
          type: ["string", "null"]
        profilePicUrl:
          type: ["string", "null"]
        socialLinks:
          $ref: "#/components/schemas/SocialLinks"

    # SONG
    Song:
      type: object
      required: [id, title, artistId, artistName, genre, isPremium, playCount, likeCount, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        artistId:
          type: integer
        artistName:
          type: string
        artistProfilePicUrl:
          type: ["string", "null"]
        genre:
          type: string
        description:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        audioUrl:
          type: string
        isPremium:
          type: boolean
        allowDownload:
          type: boolean
        playCount:
          type: integer
        likeCount:
          type: integer
        isLiked:
          type: ["boolean", "null"]
        durationSeconds:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, approved, rejected]
        createdAt:
          type: string
    SongInput:
      type: object
      required: [title, genre, audioUrl]
      properties:
        title:
          type: string
        genre:
          type: string
        description:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        audioUrl:
          type: string
        isPremium:
          type: boolean
        allowDownload:
          type: boolean
        durationSeconds:
          type: ["integer", "null"]
    SongUpdate:
      type: object
      properties:
        title:
          type: string
        genre:
          type: string
        description:
          type: ["string", "null"]
        artworkUrl:
          type: ["string", "null"]
        isPremium:
          type: boolean
        allowDownload:
          type: boolean

    # COMMENTS
    Comment:
      type: object
      required: [id, songId, userId, username, body, createdAt]
      properties:
        id:
          type: integer
        songId:
          type: integer
        userId:
          type: integer
        username:
          type: string
        profilePicUrl:
          type: ["string", "null"]
        body:
          type: string
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1

    # DISCOVER
...