# lib/api-zod/src/generated/types/comment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * DUODROP Music Streaming Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Comment {
  id: number;
  songId: number;
  userId: number;
  username: string;
  /** @nullable */
  profilePicUrl?: string | null;
  body: string;
  createdAt: string;
}